backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-genai
stripe
fpdf2
pydantic
